requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
docx2txt
faiss-cpu
PyPDF2
pdfminer.six==20221105
pdfplumber
